% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bissection.R
\name{bissection}
\alias{bissection}
\title{Bisection Method
A function to find the root of a given function within an interval using the bisection method.}
\usage{
bissection(f, a, b, MAXITERATIONS = 10000, tol = 1e-06)
}
\arguments{
\item{f}{A function for which the root is sought.}

\item{a}{The starting point of the interval.}

\item{b}{The ending point of the interval.}

\item{MAXITERATIONS}{Maximum number of iterations.}

\item{tol}{Tolerance for stopping criteria.}
}
\value{
The estimated root of the function.
}
\description{
Bisection Method
A function to find the root of a given function within an interval using the bisection method.
}
