% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton.R
\name{newton}
\alias{newton}
\title{Newton-Raphson Method}
\usage{
newton(f, x0, MAXITERATIONS = 10000, tol = 1e-06)
}
\arguments{
\item{f}{A function for which the root is sought.}

\item{x0}{The initial guess for the root.}

\item{MAXITERATIONS}{Maximum number of iterations before stopping.}

\item{tol}{The tolerance for stopping criteria. The method stops when the difference between consecutive estimates is smaller than this tolerance.}
}
\value{
A list with the root of the function, the value of the function at that root, and the number of iterations it took.
}
\description{
A function to find the root of a given function using the Newton-Raphson method.
This method uses the function values and their derivatives to approximate the root.
}
