% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secant.R
\name{secant}
\alias{secant}
\title{Secant Method}
\usage{
secant(f, x0, x1, MAXITERATIONS = 10000, tol = 1e-06)
}
\arguments{
\item{f}{A function for which the root is sought.}

\item{x0}{The first initial guess for the root.}

\item{x1}{The second initial guess for the root.}

\item{MAXITERATIONS}{Maximum number of iterations before stopping.}

\item{tol}{The tolerance for stopping criteria. The method stops when the function value is smaller than this tolerance.}
}
\value{
A list with the root of the function, the value of the function at that root, and the number of iterations it took.
}
\description{
A function to find the root of a given function using the secant method.
This method approximates the root by iteratively drawing secant lines.
}
